+++ flask_dance/exceptions.py
@@ -1,6 +1,0 @@
-# coding=utf-8
-from __future__ import unicode_literals
-
-class FlaskDanceException(Exception):
-    "Base exception class that all others in this package inherit from"
-    pass